Factor out testing and use different exit code on failure
